# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Paadi backend — auth, pots, payments, settlements",
    "title": "Paadi API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/pots": {
      "get": {
        "description": "Returns the caller's pots, filterable and paginated by query.",
        "operationId": "PotsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "properties": {
                          "collectedKobo": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "deadlineAt": {
                            "format": "date-time",
                            "nullable": true,
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "paidCount": {
                            "type": "integer"
                          },
                          "splitCount": {
                            "type": "integer"
                          },
                          "status": {
                            "enum": [
                              "draft",
                              "open",
                              "funded",
                              "settling",
                              "settled",
                              "expired",
                              "cancelled",
                              "refunding",
                              "refunded"
                            ],
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "totalKobo": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "status",
                          "totalKobo",
                          "collectedKobo",
                          "splitCount",
                          "paidCount",
                          "deadlineAt",
                          "createdAt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pots",
        "tags": [
          "pots"
        ]
      },
      "post": {
        "description": "Opens a new bill-split pot. Requires an idempotency-key header.",
        "operationId": "PotsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "attributionMode": {
                    "default": "checkout_link",
                    "enum": [
                      "checkout_link",
                      "virtual_account"
                    ],
                    "type": "string"
                  },
                  "billerCategory": {
                    "enum": [
                      "electricity",
                      "cable"
                    ],
                    "type": "string"
                  },
                  "billerCustomerId": {
                    "maxLength": 64,
                    "minLength": 1,
                    "type": "string"
                  },
                  "billerProductCode": {
                    "maxLength": 64,
                    "minLength": 1,
                    "type": "string"
                  },
                  "completionRule": {
                    "enum": [
                      "progressive",
                      "all_or_nothing"
                    ],
                    "type": "string"
                  },
                  "deadlineAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "description": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "meterType": {
                    "enum": [
                      "PREPAID",
                      "POSTPAID"
                    ],
                    "type": "string"
                  },
                  "payoutAccountId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "settlementType": {
                    "enum": [
                      "bill_payment",
                      "bank_payout",
                      "wallet"
                    ],
                    "type": "string"
                  },
                  "splitMode": {
                    "default": "weight",
                    "enum": [
                      "weight",
                      "amount",
                      "percent"
                    ],
                    "type": "string"
                  },
                  "splits": {
                    "items": {
                      "properties": {
                        "amountKobo": {
                          "exclusiveMinimum": 0,
                          "type": "integer"
                        },
                        "label": {
                          "maxLength": 80,
                          "minLength": 1,
                          "type": "string"
                        },
                        "percent": {
                          "exclusiveMinimum": 0,
                          "maximum": 100,
                          "type": "number"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "weight": {
                          "exclusiveMinimum": 0,
                          "type": "number"
                        }
                      },
                      "required": [
                        "label"
                      ],
                      "type": "object"
                    },
                    "maxItems": 50,
                    "minItems": 2,
                    "type": "array"
                  },
                  "title": {
                    "maxLength": 120,
                    "minLength": 3,
                    "type": "string"
                  },
                  "totalKobo": {
                    "exclusiveMinimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "title",
                  "totalKobo",
                  "settlementType",
                  "completionRule",
                  "splits"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "billerCategory": {
                      "enum": [
                        "electricity",
                        "cable"
                      ],
                      "nullable": true,
                      "type": "string"
                    },
                    "billerCustomerId": {
                      "nullable": true,
                      "type": "string"
                    },
                    "billerProductCode": {
                      "nullable": true,
                      "type": "string"
                    },
                    "completionRule": {
                      "enum": [
                        "progressive",
                        "all_or_nothing"
                      ],
                      "type": "string"
                    },
                    "contributions": {
                      "properties": {
                        "collectedKobo": {
                          "type": "integer"
                        },
                        "contributorCount": {
                          "type": "integer"
                        },
                        "paidSplitCount": {
                          "type": "integer"
                        },
                        "targetKobo": {
                          "type": "integer"
                        },
                        "totalSplitCount": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "contributorCount",
                        "paidSplitCount",
                        "totalSplitCount",
                        "collectedKobo",
                        "targetKobo"
                      ],
                      "type": "object"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "deadlineAt": {
                      "format": "date-time",
                      "nullable": true,
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "meterType": {
                      "enum": [
                        "PREPAID",
                        "POSTPAID"
                      ],
                      "nullable": true,
                      "type": "string"
                    },
                    "progress": {
                      "properties": {
                        "collectedKobo": {
                          "type": "integer"
                        },
                        "paidCount": {
                          "type": "integer"
                        },
                        "splitCount": {
                          "type": "integer"
                        },
                        "targetKobo": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "collectedKobo",
                        "targetKobo",
                        "paidCount",
                        "splitCount"
                      ],
                      "type": "object"
                    },
                    "settlement": {
                      "nullable": true,
                      "properties": {
                        "amountKobo": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "createdAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "failureReason": {
                          "enum": [
                            "kyc_tier_required",
                            "name_mismatch",
                            "payout_account_invalid",
                            "biller_fields_missing",
                            "gateway",
                            "provider_declined"
                          ],
                          "nullable": true,
                          "type": "string"
                        },
                        "feeKobo": {
                          "minimum": 0,
                          "type": "integer"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "nombaRef": {
                          "nullable": true,
                          "type": "string"
                        },
                        "potId": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "providerStatus": {
                          "nullable": true,
                          "type": "string"
                        },
                        "settledAt": {
                          "format": "date-time",
                          "nullable": true,
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                          ],
                          "type": "string"
                        },
                        "type": {
                          "enum": [
                            "bill_payment",
                            "bank_payout",
                            "wallet"
                          ],
                          "type": "string"
                        },
                        "vendToken": {
                          "nullable": true,
                          "type": "string"
                        },
                        "vendUnits": {
                          "nullable": true,
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "potId",
                        "type",
                        "status",
                        "amountKobo",
                        "feeKobo",
                        "vendToken",
                        "vendUnits",
                        "nombaRef",
                        "providerStatus",
                        "failureReason",
                        "settledAt",
                        "createdAt"
                      ],
                      "type": "object"
                    },
                    "settlementType": {
                      "enum": [
                        "bill_payment",
                        "bank_payout",
                        "wallet"
                      ],
                      "type": "string"
                    },
                    "splits": {
                      "items": {
                        "properties": {
                          "checkoutUrl": {
                            "nullable": true,
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "paidAt": {
                            "format": "date-time",
                            "nullable": true,
                            "type": "string"
                          },
                          "paidKobo": {
                            "type": "integer"
                          },
                          "payToken": {
                            "type": "string"
                          },
                          "shareKobo": {
                            "type": "integer"
                          },
                          "status": {
                            "enum": [
                              "pending",
                              "partially_paid",
                              "paid",
                              "overpaid",
                              "expired",
                              "refunded"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "label",
                          "shareKobo",
                          "paidKobo",
                          "status",
                          "payToken",
                          "checkoutUrl",
                          "paidAt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "status": {
                      "enum": [
                        "draft",
                        "open",
                        "funded",
                        "settling",
                        "settled",
                        "expired",
                        "cancelled",
                        "refunding",
                        "refunded"
                      ],
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "totalKobo": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "description",
                    "totalKobo",
                    "settlementType",
                    "completionRule",
                    "status",
                    "billerCategory",
                    "billerProductCode",
                    "billerCustomerId",
                    "meterType",
                    "deadlineAt",
                    "createdAt",
                    "progress",
                    "settlement",
                    "splits"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a pot",
        "tags": [
          "pots"
        ]
      }
    },
...
  },
  "servers": [],
...
}